simplify load gen code using new controller API abstraction
