--- src/Ilúvatar/iluvatar_load_gen/src/trace/controller_live.rs
@@ -1,149 +1,0 @@
-use super::{Function, TraceArgs};
-use crate::trace::{prepare_function_args, trace_utils::map_functions_to_prep, CsvInvocation};
-use crate::{
-    benchmark::BenchmarkStore,
-    trace::trace_utils::save_controller_results,
-    utils::{
-        controller_invoke, controller_prewarm, controller_register, load_benchmark_data, resolve_handles,
-        CompletedControllerInvocation, VERSION,
-    },
-};
-use anyhow::Result;
-use iluvatar_controller_library::server::controller_comm::ControllerAPIFactory;
-use iluvatar_library::types::CommunicationMethod;
-use iluvatar_library::{logging::LocalTime, transaction::gen_tid, utils::port::Port};
-use std::{
-    collections::HashMap,
-    sync::Arc,
-    time::{Duration, SystemTime},
-};
-use tokio::{runtime::Builder, task::JoinHandle};
-
-async fn controller_live_register_functions(
-    funcs: &HashMap<String, Function>,
-    host: &str,
-    port: Port,
-    benchmark: Option<&BenchmarkStore>,
-    factory: Arc<ControllerAPIFactory>,
-) -> Result<()> {
-    for (fid, func) in funcs.iter() {
-        let image = func
-            .image_name
-            .as_ref()
-            .ok_or_else(|| anyhow::anyhow!("Unable to get image name for function '{}'", fid))?;
-        println!("{}, {}", func.func_name, image);
-        let func_timings = match &func.chosen_name {
-            Some(chosen_name) => match benchmark.as_ref() {
-                Some(t) => match t.data.get(chosen_name) {
-                    Some(d) => Some(&d.resource_data),
-                    None => anyhow::bail!(format!(
-                        "Benchmark was passed but function '{}' was not present",
-                        chosen_name
-                    )),
-                },
-                None => None,
-            },
-            None => None,
-        };
-        let api = factory
-            .get_controller_api(host, port, CommunicationMethod::RPC, &gen_tid())
-            .await?;
-        let _reg_dur =
-            controller_register(&func.func_name, &VERSION, image, func.mem_mb + 50, func_timings, api).await?;
-    }
-    Ok(())
-}
-
-async fn prewarm_funcs(
-    funcs: &HashMap<String, Function>,
-    host: &str,
-    port: Port,
-    factory: Arc<ControllerAPIFactory>,
-) -> Result<()> {
-    for (fid, func) in funcs.iter() {
-        for i in 0..func.prewarms.ok_or_else(|| {
-            anyhow::anyhow!(
-                "Function '{}' did not have a prewarm value, supply one or pass a benchmark file",
-                fid
-            )
-        })? {
-            let tid = format!("{}-prewarm-{}", fid, i);
-            let api = factory
-                .get_controller_api(host, port, CommunicationMethod::RPC, &tid)
-                .await?;
-            let _reg_dur = controller_prewarm(&func.func_name, &VERSION, api, &tid).await?;
-        }
-    }
-    Ok(())
-}
-
-pub fn controller_trace_live(args: TraceArgs) -> Result<()> {
-    let mut metadata = super::load_metadata(&args.metadata_csv)?;
-    let threaded_rt = Builder::new_multi_thread().enable_all().build().unwrap();
-    let factory = ControllerAPIFactory::boxed();
-
-    map_functions_to_prep(
-        crate::utils::RunType::Live,
-        args.load_type,
-        &args.function_data,
-        &mut metadata,
-        args.prewarms,
-        &args.input_csv,
-        args.max_prewarms,
-    )?;
-    let bench_data = load_benchmark_data(&args.function_data)?;
-    threaded_rt.block_on(controller_live_register_functions(
-        &metadata,
-        &args.host,
-        args.port,
-        bench_data.as_ref(),
-        factory.clone(),
-    ))?;
-    threaded_rt.block_on(prewarm_funcs(&metadata, &args.host, args.port, factory.clone()))?;
-    let api = threaded_rt.block_on(factory.get_controller_api(
-        &args.host,
-        args.port,
-        CommunicationMethod::RPC,
-        &gen_tid(),
-    ))?;
-
-    let mut trace_rdr = csv::Reader::from_path(&args.input_csv)?;
-    let mut handles: Vec<JoinHandle<Result<CompletedControllerInvocation>>> = Vec::new();
-    let clock = Arc::new(LocalTime::new(&gen_tid())?);
-
-    println!("starting live trace run");
-
-    let start = SystemTime::now();
-    for result in trace_rdr.deserialize() {
-        let invocation: CsvInvocation = result?;
-        let func = metadata.get(&invocation.func_name).ok_or_else(|| {
-            anyhow::anyhow!(
-                "Invocation had function name '{}' that wasn't in metadata",
-                invocation.func_name
-            )
-        })?;
-        let func_args = prepare_function_args(func, args.load_type);
-
-        let clk_cln = clock.clone();
-        let api_cln = api.clone();
-        let f_c = func.func_name.clone();
-        loop {
-            match start.elapsed() {
-                Ok(t) => {
-                    let ms = t.as_millis() as u64;
-                    if ms >= invocation.invoke_time_ms {
-                        break;
-                    }
-                    std::thread::sleep(Duration::from_millis(ms / 2));
-                }
-                Err(_) => (),
-            }
-        }
-        handles.push(
-            threaded_rt
-                .spawn(async move { controller_invoke(&f_c, &VERSION, Some(func_args), clk_cln, api_cln).await }),
-        );
-    }
-    let results = resolve_handles(&threaded_rt, handles, crate::utils::ErrorHandling::Print)?;
-    save_controller_results(results, &args)
-}
